pyproject.toml
 [project]
 name = "arcastone"
 version = "0.1.0"
 description = "ArcaStone - Offline-first desktop PDF index and search (PySide6)"
 readme = "README.md"
 requires-python = ">=3.11"
 dependencies = [
   "PySide6==6.7.2",
   "blake3==0.4.1",
   "pypdf==5.1.0",
   "faiss-cpu==1.8.0.post1",
   "sentence-transformers==3.0.1",
   "numpy==1.26.4",
   "pandas==2.2.2",
   "sqlalchemy==2.0.32",
 ]

 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
 build-backend = "setuptools.build_meta"

 [tool.pytest.ini_options]
 addopts = "-q"


